requirements.txt
python-nmap
python-whois
tqdm
pyfiglet
requests